+++ CHANGELOG.md
@@ -22,7 +22,7 @@ Please file a bug if you notice a violation of semantic versioning.
 
 ### Added
 
-- Add `auth-sanitizer` runtime dependency to redact sensitive LDAP adaptor inspection output
+- Add `auth-sanitizer` runtime dependency to redact sensitive LDAP adaptor inspection output without defining top-level `Auth`
 
 ### Changed